refactor: 경험 상세 폼 상태 로직 분리
